allow file passthrough
