+++ src/tuitka/assets/logo/nuitka.png
