+++ src/tuitka/assets/style_inline_app.tcss
@@ -1,0 +1,21 @@
+#image {
+    text-align: center;
+    content-align: center middle;
+}
+
+#website-link {
+    text-align: center;
+    margin: 1 0;
+}
+
+#terminal-container {
+    margin: 0 2 1 2;
+    height: 1fr;
+    align: center middle;
+}
+
+#compilation_terminal {
+    height: 1fr;
+    background: $surface-lighten-3;
+    color: $text;
+}
