+++ src/tuitka/constants.py
@@ -1,3 +1,7 @@
+import sys
+
+PYTHON_VERSION = f"{sys.version_info.major}.{sys.version_info.minor}"
+
 sss_snek = r"""
   ____                  
      / . .\
