+++ src/tuitka/widgets/modals/__init__.py
@@ -1,0 +1,23 @@
+from .compilation import CompilationScreen
+from .file_dialog import FileDialogScreen
+from .settings import NuitkaSettingsScreen
+from .settings_widgets import (
+    ModalBoolFlag,
+    ModalRadioFlag,
+    ModalSelectionFlag,
+    ModalStringFlag,
+)
+from .splash import SplashScreen
+from .support import SupportNuitkaModal
+
+__all__ = [
+    "CompilationScreen",
+    "FileDialogScreen",
+    "ModalBoolFlag",
+    "ModalRadioFlag",
+    "ModalSelectionFlag",
+    "ModalStringFlag",
+    "NuitkaSettingsScreen",
+    "SplashScreen",
+    "SupportNuitkaModal",
+]
